api/requirements.txt
aioamqp==0.15.0
alembic==1.12.0
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
coverage==5.5
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==2.0.0
Django==4.2.2
docker==6.1.3
duplicity==0.8.21
Faker==14.0.0
fasteners==0.14.1
filelock==3.12.2
Flask==2.3.3
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==0.18.2
greenlet==2.0.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pamqp==3.2.1
paramiko==2.9.3
pexpect==4.8.0
pika==1.3.2
Pillow==9.0.1
platformdirs==3.10.0
protobuf==3.12.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
usb-creator==0.3.7
virtualenv==20.24.3
wadllib==1.3.6
websocket-client==1.6.2
Werkzeug==3.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
